Admin screen updates (#7)
* fix: bug fixes

+++ pages/api/products/[id].js
@@ -5,7 +5,7 @@ export default async (req, res) => {
     const { userId, categoryId, groupId, product } = req.body;
     try {
         if (req.method === 'PUT') {
-            await db.collection(`products`).doc(id).update({
+            await db.collection(`users/${userId}/products`).doc(id).update({
                 ...product,
                 groupId,
                 categoryId,
@@ -13,7 +13,7 @@ export default async (req, res) => {
                 updated: new Date().toISOString(),
             });
         } else if (req.method === 'DELETE') {
-            await db.collection(`products`).doc(id).delete();
+            await db.collection(`users/${userId}/products`).doc(id).delete();
         }
         res.status(200).end();
     } catch (e) {
